gen: simplify the config format for Go files
Instead of requiring a complete config inside a comment, extract only the
gen.Enum value. This also makes it possible to support multiple such comment
blocks in a given file, so multiple types can be defined separately.

Rename the tag "enumgen:type", and extract the type name from the remainder of
that line.